Finally cracked live updating!!!
There were issues with harvested tweet content never seeming to get into the arrays
THis was because the arrays were being appended with the whole list and getting bigger and bigger. This was causing degrading performance over time, AND the repetition meant the words being picked were biased by the duplication.